requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
pydantic
openai
pyyaml